requirements.txt
# Core Dependencies
openai-whisper>=20231117
gradio>=4.0.0
ollama>=0.1.0
torch>=2.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0

# Audio Processing
soundfile>=0.12.0
sounddevice>=0.4.6
pyaudiowpatch>=0.0.7  # Windows WASAPI loopback

# Speaker Diarization
pyannote.audio>=3.1.0
speechbrain>=1.0.0

# System Tray
pystray>=0.19.0
Pillow>=10.0.0

# UI/UX
tqdm>=4.65.0

# Optional: Better audio processing
# librosa>=0.10.0

# Development
pytest>=7.0.0
black>=23.0.0
